Refactor DSL into subpackages and rename rules
